Move gcode/stl loading code into ModelLoader
